requirements.txt
## --------------------------------------------------
## Flask and API Configuration
## --------------------------------------------------
python-dotenv
flask
flask-cors
pyjwt[crypto]
## --------------------------------------------------
## Data Preprocessor and Postprocessor
## --------------------------------------------------
numpy
## --------------------------------------------------
## Model Wrapper
## --------------------------------------------------
scikit-learn
joblib
